site/backend/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
redis==3.5.2
requests==2.23.0
rq==1.4.1
six==1.14.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
urllib3==1.25.9
Werkzeug==1.0.1
